Cmake + Visual Studio
Ok so now the cmake file works nicely with visual studio and windows.
it assumes that you have the SDL stuff, or that it's in a folder named
SDL in the repo. Which I have removed from this commit for the sake of keeping
this repository clean.

Now the game needs to be made... Quickly...