Renderers
Added a bunch of rendering stuffs.
Rendering is based entirely on data inside ofassets now. So if you want to render a model, it'll look at the model file and load all the appropriate textures and stuff based on the file name alone.

there are now 3 types of renderer too... And sdl renderer, an openglSdl renderer and a windows Gl renderer. There are memory leaks with the openglsdl renderer in windows for some reason.

added bezier curves. added bad fuzzy logic. (fuzzy logic sucks anyways). Made the camera a bit better. Added some basic physics functionality. Will be working harder on that. Fixed some stuff to do with types.

Lots of stuff and that i can't remember probably.