Closer, messed up selects.
This is starting to look more right. I hadn't set good timeouts for
select and it was causing a lot of issues. I think this needs to be
tuned. I also turned off nagles algorithm because I was worried that
the client wasn't recieving the final bytes of a file. Doesn't seem to
be the case. Definitely something wrong with file send vs
content-length.

Will setup packet sniffers at some point to see what actually goes
down the pipe.