Stalls fixed, but video transfer isn't right.
So, on mozilla firefox it works (kinda) but on safari, not at all.
The most it's buffered is 3 seconds of the 10 second video.

Apart from the fixed stalls I also added a landing page for
rc_sample. Links to nipple.hml and movie.html, so you can navigate
instead of typing the urls repeatedly. Works well.